backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
redis[hiredis]==5.0.1

# TimescaleDB
psycopg2-binary==2.9.9

# LLM Providers
openai==1.3.7
google-generativeai>=0.8.0
tiktoken==0.5.2

# Embeddings
numpy==1.26.2

# NLP and PII Detection
spacy==3.7.2
# Note: spaCy models are downloaded in Dockerfile via download_models.py script

# Security
bcrypt==4.1.1
cryptography==41.0.7

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Configuration
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
mypy==1.7.1
ruff==0.1.6

